// src/api/generated/models/activityLogListResponseDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Flowforge API
 * The Flowforge workflow orchestration engine API documentation
 * OpenAPI spec version: 1.0
 */
import type { ActivityLogItemDto } from './activityLogItemDto';
import type { PaginationMetaDto } from './paginationMetaDto';

export interface ActivityLogListResponseDto {
  message: string;
  data: ActivityLogItemDto[];
  meta: PaginationMetaDto;
}